+++ core/delete_volume_group_backup_request_response.go
@@ -1,25 +1,23 @@
 // Copyright (c) 2016, 2018, Oracle and/or its affiliates. All rights reserved.
 // Code generated. DO NOT EDIT.
 
-package database
+package core
 
 import (
 	"github.com/oracle/oci-go-sdk/common"
 	"net/http"
 )
 
-// LaunchDbSystemRequest wrapper for the LaunchDbSystem operation
-type LaunchDbSystemRequest struct {
+// DeleteVolumeGroupBackupRequest wrapper for the DeleteVolumeGroupBackup operation
+type DeleteVolumeGroupBackupRequest struct {
 
-	// Request to launch a DB System.
-	LaunchDbSystemDetails LaunchDbSystemBase `contributesTo:"body"`
+	// The Oracle Cloud ID (OCID) that uniquely identifies the volume group backup.
+	VolumeGroupBackupId *string `mandatory:"true" contributesTo:"path" name:"volumeGroupBackupId"`
 
-	// A token that uniquely identifies a request so it can be retried in case of a timeout or
-	// server error without risk of executing that same action again. Retry tokens expire after 24
-	// hours, but can be invalidated before then due to conflicting operations (for example, if a resource
-	// has been deleted and purged from the system, then a retry of the original creation request
-	// may be rejected).
-	OpcRetryToken *string `mandatory:"false" contributesTo:"header" name:"opc-retry-token"`
+	// For optimistic concurrency control. In the PUT or DELETE call for a resource, set the `if-match`
+	// parameter to the value of the etag from a previous GET or POST response for that resource.  The resource
+	// will be updated or deleted only if the etag you provide matches the resource's current etag value.
+	IfMatch *string `mandatory:"false" contributesTo:"header" name:"if-match"`
 
 	// Unique Oracle-assigned identifier for the request.
 	// If you need to contact Oracle about a particular request, please provide the request ID.
@@ -30,42 +28,36 @@ type LaunchDbSystemRequest struct {
 	RequestMetadata common.RequestMetadata
 }
 
-func (request LaunchDbSystemRequest) String() string {
+func (request DeleteVolumeGroupBackupRequest) String() string {
 	return common.PointerString(request)
 }
 
 // HTTPRequest implements the OCIRequest interface
-func (request LaunchDbSystemRequest) HTTPRequest(method, path string) (http.Request, error) {
+func (request DeleteVolumeGroupBackupRequest) HTTPRequest(method, path string) (http.Request, error) {
 	return common.MakeDefaultHTTPRequestWithTaggedStruct(method, path, request)
 }
 
 // RetryPolicy implements the OCIRetryableRequest interface. This retrieves the specified retry policy.
-func (request LaunchDbSystemRequest) RetryPolicy() *common.RetryPolicy {
+func (request DeleteVolumeGroupBackupRequest) RetryPolicy() *common.RetryPolicy {
 	return request.RequestMetadata.RetryPolicy
 }
 
-// LaunchDbSystemResponse wrapper for the LaunchDbSystem operation
-type LaunchDbSystemResponse struct {
+// DeleteVolumeGroupBackupResponse wrapper for the DeleteVolumeGroupBackup operation
+type DeleteVolumeGroupBackupResponse struct {
 
 	// The underlying http response
 	RawResponse *http.Response
-
-	// The DbSystem instance
-	DbSystem `presentIn:"body"`
-
-	// For optimistic concurrency control. See `if-match`.
-	Etag *string `presentIn:"header" name:"etag"`
 
 	// Unique Oracle-assigned identifier for the request. If you need to contact Oracle about
 	// a particular request, please provide the request ID.
 	OpcRequestId *string `presentIn:"header" name:"opc-request-id"`
 }
 
-func (response LaunchDbSystemResponse) String() string {
+func (response DeleteVolumeGroupBackupResponse) String() string {
 	return common.PointerString(response)
 }
 
 // HTTPResponse implements the OCIResponse interface
-func (response LaunchDbSystemResponse) HTTPResponse() *http.Response {
+func (response DeleteVolumeGroupBackupResponse) HTTPResponse() *http.Response {
 	return response.RawResponse
 }