+++ core/get_volume_group_request_response.go
@@ -1,25 +1,18 @@
 // Copyright (c) 2016, 2018, Oracle and/or its affiliates. All rights reserved.
 // Code generated. DO NOT EDIT.
 
-package database
+package core
 
 import (
 	"github.com/oracle/oci-go-sdk/common"
 	"net/http"
 )
 
-// LaunchDbSystemRequest wrapper for the LaunchDbSystem operation
-type LaunchDbSystemRequest struct {
+// GetVolumeGroupRequest wrapper for the GetVolumeGroup operation
+type GetVolumeGroupRequest struct {
 
-	// Request to launch a DB System.
-	LaunchDbSystemDetails LaunchDbSystemBase `contributesTo:"body"`
-
-	// A token that uniquely identifies a request so it can be retried in case of a timeout or
-	// server error without risk of executing that same action again. Retry tokens expire after 24
-	// hours, but can be invalidated before then due to conflicting operations (for example, if a resource
-	// has been deleted and purged from the system, then a retry of the original creation request
-	// may be rejected).
-	OpcRetryToken *string `mandatory:"false" contributesTo:"header" name:"opc-retry-token"`
+	// The Oracle Cloud ID (OCID) that uniquely identifies the volume group.
+	VolumeGroupId *string `mandatory:"true" contributesTo:"path" name:"volumeGroupId"`
 
 	// Unique Oracle-assigned identifier for the request.
 	// If you need to contact Oracle about a particular request, please provide the request ID.
@@ -30,28 +23,28 @@ type LaunchDbSystemRequest struct {
 	RequestMetadata common.RequestMetadata
 }
 
-func (request LaunchDbSystemRequest) String() string {
+func (request GetVolumeGroupRequest) String() string {
 	return common.PointerString(request)
 }
 
 // HTTPRequest implements the OCIRequest interface
-func (request LaunchDbSystemRequest) HTTPRequest(method, path string) (http.Request, error) {
+func (request GetVolumeGroupRequest) HTTPRequest(method, path string) (http.Request, error) {
 	return common.MakeDefaultHTTPRequestWithTaggedStruct(method, path, request)
 }
 
 // RetryPolicy implements the OCIRetryableRequest interface. This retrieves the specified retry policy.
-func (request LaunchDbSystemRequest) RetryPolicy() *common.RetryPolicy {
+func (request GetVolumeGroupRequest) RetryPolicy() *common.RetryPolicy {
 	return request.RequestMetadata.RetryPolicy
 }
 
-// LaunchDbSystemResponse wrapper for the LaunchDbSystem operation
-type LaunchDbSystemResponse struct {
+// GetVolumeGroupResponse wrapper for the GetVolumeGroup operation
+type GetVolumeGroupResponse struct {
 
 	// The underlying http response
 	RawResponse *http.Response
 
-	// The DbSystem instance
-	DbSystem `presentIn:"body"`
+	// The VolumeGroup instance
+	VolumeGroup `presentIn:"body"`
 
 	// For optimistic concurrency control. See `if-match`.
 	Etag *string `presentIn:"header" name:"etag"`
@@ -61,11 +54,11 @@ type LaunchDbSystemResponse struct {
 	OpcRequestId *string `presentIn:"header" name:"opc-request-id"`
 }
 
-func (response LaunchDbSystemResponse) String() string {
+func (response GetVolumeGroupResponse) String() string {
 	return common.PointerString(response)
 }
 
 // HTTPResponse implements the OCIResponse interface
-func (response LaunchDbSystemResponse) HTTPResponse() *http.Response {
+func (response GetVolumeGroupResponse) HTTPResponse() *http.Response {
 	return response.RawResponse
 }